refactor: moved data gen to context file
